pyproject.toml
[tool.poetry]
name = "goinsightmanager"
version = "0.1.0"
description = ""
authors = ["takeru.furuse <takeru.furuse@konicaminolta.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.0"
django-stubs = "^1.15.0"
django-mptt = "^0.14.0"
django-extensions = "^3.2.1"
django-environ = "^0.10.0"
django-nested-admin = "^4.0.2"
ulid-py = "^1.1.0"
psycopg2 = "^2.9.6"
python-dateutil = "^2.8.2"

[tool.poetry.group.tool.dependencies]
pre-commit = "^3.1.1"
flake8 = "^6.0.0"
flake8-print = "^5.0.0"
flake8-pyi = "^23.3.1"
black = "^23.1.0"
isort = "^5.12.0"
mypy = "^1.1.1"
werkzeug = "^2.2.3"
ipython = "^8.12.0"

[tool.poetry.group.dev.dependencies]
django-stubs = {extras = ["compatible-mypy"], version = "^1.16.0"}
django-debug-toolbar = "^4.0.0"
types-python-dateutil = "^2.8.19.12"

[tool.black]
line-length = 120
include = '\.pyi?$'
extend-exclude = '''
/(
    \.pytest_cache |
    \.mypy_cache |
    \.venv |
    venv |
    migrations
)/
'''

[tool.isort]
profile = "black"
skip = "migrations"

[tool.mypy]
mypy_path = "./src"
ignore_missing_imports = true
show_error_context = true
pretty = true
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "conf.settings.base"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
